// frontend/src/app/boxeador/[id]/types.ts
export interface Boxer {
    nombre: string;
    apellidos: string;
    fecha_nacimiento: string;
    sexo: string;
    dni: string;
    numero_licencia: string;
    nacionalidad: string;
    provincia: string;
    localidad: string;
    debutaje: string;
    carrera: string;
    peso: string;
    divicion: string; // yase que esta mal escrito
    status : string;
    combates: string;
    rounds: string;
    KOs: string;
    altura: string;
    alcance: string;
    stance: string;
    victorias: string;
    derrotas: string;
    empates: string;
    sin_definir: string;
    foto: string;
}